Second Commit - Changed some views

+++ app/src/main/java/com/example/JobSpot/activity_removesupply.kt
@@ -4,6 +4,7 @@ import android.content.Intent
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.widget.Toast
+import androidx.appcompat.app.AlertDialog
 import com.example.JobSpot.databinding.ActivityRemovesupplyBinding
 import com.google.firebase.database.DatabaseReference
 import com.google.firebase.database.FirebaseDatabase
@@ -24,8 +25,19 @@ class activity_removesupply : AppCompatActivity() {
 
         binding.btnRemove.setOnClickListener {
             var deleteSupplier = binding.edtDelete.text.toString()
-            if(deleteSupplier.isNotEmpty())
-                deleteData(deleteSupplier)
+            if(deleteSupplier.isNotEmpty()) {
+                val builder = AlertDialog.Builder(this)
+                builder.setTitle("Delete Job")
+                    .setMessage("Are you sure you want to delete this job/task?")
+                    .setPositiveButton("Yes") { _, _ ->
+                        // code to delete record
+                        deleteData(deleteSupplier)
+                    }
+                    .setNegativeButton("No") { dialog, _ ->
+                        dialog.dismiss()
+                    }
+                    .create().show()
+            }
             else
                 Toast.makeText(this,"Please enter JobID",Toast.LENGTH_SHORT).show()
         }
